added reply action

+++ database/migrations/2020_12_18_000002_create_chapar_subjects_table.php
@@ -21,9 +21,14 @@ class CreateChaparSubjectsTable extends Migration
         });
 
         Schema::table('chapar_letters', function (Blueprint $table) {
-            $table->foreignId('subject_id'); 
-            $table->dropColumn('subject');   
+            $table->foreignId('subject_id')->nullable();     
         });
+
+        if(Schema::hasColumn('chapar_letters', 'subject')) { 
+            Schema::table('chapar_letters', function (Blueprint $table) { 
+                $table->dropColumn('subject'); 
+            });
+        }
     }
 
     /**
